redesign: records page table
Context:
Previously the method of displayign the table on mobile was to stack the
<td>s in a column, with headings to the side. While this was /usable/
it used a lot more vertical space than it needed to.

Additionally, the tooltips for PO numbers and Product lists didn't look
great and I wanted to move those to my new "modal".

Goals:
> Mobile: display as cards, to make more efficient use of space
> Desktop: stretch the table, to fill the space, and maybe improve
  formatting
> Middling screen widths: apply horizontal-scrolling to the table,
  instead of jumping to cards prematurely (while horizontal scrolling
  is annoying, so are cards, so only swap to those when there'd be a
  /lot/ of horizontal scrolling)

Changes:
> Table now switches to cards on mobile
> Added horizontal scrolling, with conditional shadows at the sides
> The full list of products now opens in a modal, instead of a
  tooltip
> Country is now displayed as a flag, where hovering over it opens a
  tooltip with the country's name
> Some buttons replaced with icons (i.e. link to job page, button to
  open products modal, add to todo list) to save space
> CSS for tooltip labels on buttons/icons has been split off, for
  shared usage
> Purchase order modal removed, replaced with a comma-separated string
  which will gain a scrollbar if it won't fit (*)
> Added handling for long company names
> models.py and util.py updated to "break up" the information for the
  products modal and the country flags to support the new formatting

(*) PO numbers are important info for identifying the correct order -
the main purpose of the Records page - so we don't want them
squirrelled away in a modal just because there's a second one. Better
to at least see the first and side-scroll if needed. This might be a
pain if there were ever many, many POs on a single job, but that should
be rare.

Naughty unrelated changes:
> Tidied records_filter.js and todo_comments.js: removed commented out
  code and console.log()s
> Bug fix: the tooltips on the todo list's buttons were activing
  ":hover" effects when invisible, with a possible consequence being
  removing a job from the todo list when you were trying to click the
  link to the job page. Turned off pointer-events on the tooltips to
  address this