format: center remaining Job sections
On wider monitors, many of the Job page's sections were aligned left.
The recently adjusted Comments section broke the mould, being aligned
to the center; it was joined by the Price Check section.

For consistency (and to avoid everything being squeezed up to the left
when you're looking in the middle of the screen) this commit centers
the remaining sections, except Summary (which makes better use of the
horizontal space)

* Centered Job > Documents
* Adjusted the Job > Items "details window" to always appear on wider
  screens, next to the table
* Centered Job > Items (+ Details)
* Grouped General Details and Purchase Orders (PO) onto a single page:
  they both looked a bit lost, floating in the middle of the screen.
* Adjusted formatting of General Details and PO sections to reflect
  becoming panels

Side-fixes:
* Fixed bug that prevented users form adding a new PO
* Restored row banding formatting to Records table