fix: Job Summary failure and status codes
Two small bugfixes.

Bug #1: Job > Summary wouldn't build for new Jobs
Turns out this was because the code wasn't setup to handle new Jobs,
where there weren't any comments. Added a suitable check.

Bug #2: Incorrect status codes on new Jobs
When a new Job was created, despite having no WO or OC (or anything to
put on a WO or an OC) both were showing up as "ok".

Added a line to check that there was at least 1 item on an issued
document before declaring it to be "ok".