format: Safari bug with comments filter; + misc
Bug
In Safari, on the first render of the comments filter, the heights of
the filter labels are huge. They overlap one another, exceed vertical
space and activate scrolling. It looked like the third label in each
radio - "Both", which appears in all three radios and is the default
selection in all three - had merged into a single element, and/or were
overlaid.

Behaviour
* Always occurs when you first open the filter popout, regardless of
  phone orientation
* Rotating the phone to the other orientation while the filter is open
  fixes it
* Fix persists if you subsequently rotate the phone back to the original
  orientation
* Selecting a different option in "radio group A" fixes the height in
  radio group A, but not in the other two radio groups
* Fix persists if you subsequently reselect the original option in radio
  group A
* Fix also persists if you select another option in all three radios,
  then return all three radios to the default option

Attempt #1 fixes
* Added grid-template-rows in case its absence was upsetting Safari
* Added a max-height to the labels

---
Misc small updates:
* Updated interactive elements to have pointer cursors
* Removed some outdated part number formatting that had evaded an
  earlier purge
* Docitems now apply the correct formatting when 0 are assigned to a
  document